Added the data / information for all the elements in the periodic table
In the last commit where we just added the main.py file, but included just a few dummy data of the elements of the periodic table. But, in this commit, we have added all the elements of the periodic table into the main.py file. The information for the periodic table elements are stored in the main.py file using the lists and dictionaries.